% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkrer.R
\docType{data}
\name{lkrer}
\alias{lkrer}
\title{Daily Indicative Exchange Rates (LKR per world currencies)}
\format{
A tsibble with 179419 rows and 3 variables:
\describe{
\item{Currency}{currency name}
\item{Date}{date}
\item{Exchange.Rate}{Indicative exchange rate}
}
}
\source{
Central Bank of Sri Lanka
}
\usage{
lkrer
}
\description{
This dataset contains daily indicative exchange rates LKR per world currencies
}
\examples{
head(lkrer)
}
\keyword{datasets}
